update: success passing data from CTRLR -> VIEW
NOTE: THIS IS A HACKY SOLUTION, since modifying the controller form the
frontend and generating new controller updates the existing code.

This was the solution reached in one day, and hopefully, there exists a
more elegant solution for this.

right now, a dictionary object is injected in the generated controller
which will hold the result from the API.

this is then used in the view to display the data.